refactor: Add ChangeWallpaper component for customizing wallpaper in SearchPage and globally
