+++ src/components/Settings/components/SearchPage.vue
@@ -1,11 +1,10 @@
 <script lang="ts" setup>
-import Input from '~/components/Input.vue'
 import Radio from '~/components/Radio.vue'
-import Slider from '~/components/Slider.vue'
 import Tooltip from '~/components/Tooltip.vue'
-import { SEARCH_BAR_CHARACTERS, WALLPAPERS } from '~/constants/imgs'
+import { SEARCH_BAR_CHARACTERS } from '~/constants/imgs'
 import { settings } from '~/logic'
 
+import ChangeWallpaper from './ChangeWallpaper.vue'
 import SettingsItem from './SettingsItem.vue'
 import SettingsItemGroup from './SettingsItemGroup.vue'
 
@@ -18,10 +17,6 @@ watch(() => settings.value.individuallySetSearchPageWallpaper, (newValue) => {
 
 function changeSearchBarFocusCharacter(url: string) {
   settings.value.searchPageSearchBarFocusCharacter = url
-}
-
-function changeWallpaper(url: string) {
-  settings.value.searchPageWallpaper = url
 }
 </script>
 
@@ -113,93 +108,7 @@ function changeWallpaper(url: string) {
 
         <Radio v-model="settings.individuallySetSearchPageWallpaper" />
       </SettingsItem>
-      <template v-if="settings.individuallySetSearchPageWallpaper">
-        <SettingsItem :title="$t('settings.wallpaper_mode')" :desc="$t('settings.wallpaper_mode_desc')">
-          <div w-full flex rounded="$bew-radius" bg="$bew-fill-1" p-1>
-            <div
-              flex-1 py-1 cursor-pointer text-center rounded="$bew-radius"
-              :style="{
-                background: settings.searchPageWallpaperMode === 'buildIn' ? 'var(--bew-theme-color)' : '',
-                color: settings.searchPageWallpaperMode === 'buildIn' ? 'white' : '',
-              }"
-              @click="settings.searchPageWallpaperMode = 'buildIn'"
-            >
-              {{ $t('settings.wallpaper_mode_opt.build_in') }}
-            </div>
-            <div
-              flex-1 py-1 cursor-pointer text-center rounded="$bew-radius"
-              :style="{
-                background: settings.searchPageWallpaperMode === 'byUrl' ? 'var(--bew-theme-color)' : '',
-                color: settings.searchPageWallpaperMode === 'byUrl' ? 'white' : '',
-              }"
-              @click="settings.searchPageWallpaperMode = 'byUrl'"
-            >
-              {{ $t('settings.wallpaper_mode_opt.by_url') }}
-            </div>
-          </div>
-        </SettingsItem>
-
-        <SettingsItem v-if="settings.searchPageWallpaperMode === 'buildIn'" :title="$t('settings.choose_ur_wallpaper')" next-line>
-          <div grid="~ xl:cols-5 lg:cols-4 cols-3 gap-4">
-            <picture
-              aspect-video bg="$bew-fill-1" rounded="$bew-radius" overflow-hidden
-              un-border="4 transparent" cursor-pointer
-              grid place-items-center
-              :class="{ 'selected-wallpaper': settings.searchPageWallpaper === '' }"
-              @click="changeWallpaper('')"
-            >
-              <div i-tabler:photo-off text="3xl $bew-text-3" />
-            </picture>
-            <Tooltip v-for="item in WALLPAPERS" :key="item.url" placement="top" :content="item.name" aspect-video>
-              <picture
-                aspect-video bg="$bew-fill-1" rounded="$bew-radius" overflow-hidden
-                un-border="4 transparent" w-full
-                :class="{ 'selected-wallpaper': settings.searchPageWallpaper === item.url }"
-                @click="changeWallpaper(item.url)"
-              >
-                <img :src="item.thumbnail" alt="" w-full h-full object-cover>
-              </picture>
-            </Tooltip>
-          </div>
-        </SettingsItem>
-        <SettingsItem v-else :title="$t('settings.image_url')" next-line>
-          <div flex items-center gap-4>
-            <picture
-              aspect-video bg="$bew-fill-1" rounded="$bew-radius" overflow-hidden
-              un-border="4 transparent" cursor-pointer shrink-0
-              w="xl:1/5 lg:1/4 md:1/3"
-            >
-              <img
-                v-if="settings.searchPageWallpaper" :src="settings.searchPageWallpaper" alt="" w-full h-full
-                object-cover onerror="this.style.display='none'; this.onerror=null;"
-              >
-            </picture>
-            <div>
-              <Input v-model="settings.searchPageWallpaper" w-full />
-              <p color="sm $bew-text-3" mt-2>
-                {{ $t('settings.image_url_hint') }}
-              </p>
-            </div>
-          </div>
-        </SettingsItem>
-
-        <SettingsItem :title="$t('settings.enable_wallpaper_masking')">
-          <template #desc>
-            <span color="$bew-warning-color">{{ $t('common.performance_impact_warn') }}</span>
-          </template>
-
-          <Radio v-model="settings.searchPageEnableWallpaperMasking" />
-        </SettingsItem>
-        <SettingsItem v-if="settings.searchPageEnableWallpaperMasking" :title="$t('settings.wallpaper_mask_opacity')">
-          <Slider v-model="settings.searchPageWallpaperMaskOpacity" :label="`${settings.searchPageWallpaperMaskOpacity ?? 0}%`" />
-        </SettingsItem>
-        <SettingsItem v-if="settings.searchPageEnableWallpaperMasking" :title="$t('settings.wallpaper_blur_intensity')">
-          <template #desc>
-            <span color="$bew-warning-color">{{ $t('common.performance_impact_warn') }}</span>
-          </template>
-          <Slider v-model="settings.searchPageWallpaperBlurIntensity" :min="0" :max="60" :label="`${settings.searchPageWallpaperBlurIntensity ?? 0}px`" />
-        </SettingsItem>
-      </template>
+      <ChangeWallpaper v-if="settings.individuallySetSearchPageWallpaper" type="searchPage" />
     </SettingsItemGroup>
   </div>
 </template>
